+++ tests/testthat/test-worcs_project.R
@@ -21,6 +21,8 @@ test_that("worcs project can be generated", {
   expect_true(file.exists("worcs_project/.worcs"))
   expect_true(file.exists("worcs_project/LICENSE"))
   expect_true(file.exists("worcs_project/README.md"))
+  readme_contents <- readLines("worcs_project/README.md", encoding = "UTF-8")
+  expect_true(any(readme_contents == paste0("You can load this project in RStudio by opening the file called '", "worcs_project.Rproj", "'.")))
   expect_true(file.exists("worcs_project/preregistration.Rmd"))
   expect_true(file.exists("worcs_project/manuscript/manuscript.Rmd"))
   expect_error(git_status("worcs_project"), NA)